fixed favorite page
